TCF Agent: Cosmetic: fix default name of Eclipse projects in .project
The TCF agent .project files still specified "org.eclipse.tm.tcf" as the
namespace. I'm replacing this with the newer "org.eclipse.tcf" prefix.
This field is only used when importing a project into Eclipse for the
first time - and it's only the default name so no impact other than
cosmetic is expected.