Enable initialization output on stdout and stderr in daemon mode and add daemon support on windows
Prior to this change it was not possible to get the output from the -S option in daemon mode.  This made it hard for clients to start the TCF agent and know 1) which port is used, 2) when the started agent is ready for incoming connections and 3) if the launch was successful or if it failed.