Bug 408260 - Chained tcf servers do not report correct list of services
When several TCF servers are chained between a client and an agent (using Locator.redirect command), the services list of the TCF chain from the client perspective is not correctly consolidated and reported.
After the server is redirected to another TCF entity, it intercepts all Locator.redirect commands from the client to update its internal list of services.