Add protocol_send_command_with_progress() to support handling of progress messages
Prior to this change the reply handler was called for both progress
and reply messages.  However, there was no flag to differentiate
progress messages from reply messages, without interpreting the
message body.

With this change the reply handler is no longer called for progress
messages.  This is not 100% backwards compatible.  At this time we are
not aware of any service using progress messages, so it will hopefully
not affect anybody in a negative way.