Bug 422786: Fix get_isa support
This commit fixes get_isa support within Disassembly service to be ACPM compliant. This is required if Disassembly and RunControl services are implemented by two different TCF peers.
If Disassembly is implemented by the value-add, it may send a "RunControl getISA" query to the TCF agent if symbols are not available.

Signed-off-by: jguiban <jerome.guiban@windriver.com>