TCF Agent: Allow creating a proxy between two connected channels
Allow creating a proxy between two already connected channels; before
this change it was only possible to create a proxy between a connected
channel and a yet to connect channel. This can be used to implement
a bridge between two TCF peers when none of them accept incoming
connections (behind a firewall).