TCF Agent: Added PortForward and PortServer services
Added support for PortForward & PortServer services.
PortForward service allows opening a tcp/udp/serial port on the
remote peer and remote access the port input/output using a TCF
stream.

The PortServer service is built on top of PortForward service
and allows opening a local TCP/UDP port and redirect it to
a serial/udp/tcp port of the remote peer.

A PortForward service proxy library is also provided and provides
an API to locally create a PortServer.