TCF Agent: allow for a case-insensitive PathMap
Use-case: source and destination paths coming from a case-insensitive FS
(like the default on Windows). The user *might* provide the path map
rules ignoring the case, since they would still be "correct" on both the
source and the destination file systems.

The default behavior of the PathMap service is still case sensitive.
This is just to provide *choice*.

Change-Id: Ib9eebe9ea559008e728ace51dda9c4675b0260fe
Signed-off-by: Mircea Gherzan <mircea.gherzan@intel.com>